+++ src/pymetropolis/metro_spatial/simulation_area/file.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from pymetropolis.metro_pipeline.file import MetroGeoDataFrameFile
 
 if TYPE_CHECKING:
-    from shapely.geometry import Polygon
+    from shapely.geometry import MultiPolygon, Polygon
 
 
 class SimulationAreaFile(MetroGeoDataFrameFile):
@@ -13,18 +13,18 @@ class SimulationAreaFile(MetroGeoDataFrameFile):
     description = "Single-feature file with the geometry of the simulation area."
     max_rows = 1
 
-    def get_area(self) -> Polygon:
+    def get_area(self) -> Polygon | MultiPolygon:
         """Returns the simulation area as a Polygon.
 
         If the file does not exist, raises an error."""
-        from shapely.geometry import Polygon
+        from shapely.geometry import MultiPolygon, Polygon
 
         gdf = self.read()
         area = gdf["geometry"].iloc[0]
-        assert isinstance(area, Polygon)
+        assert isinstance(area, Polygon) or isinstance(area, MultiPolygon)
         return area
 
-    def get_area_opt(self) -> Polygon | None:
+    def get_area_opt(self) -> Polygon | MultiPolygon | None:
         """Returns the simulation area as a Polygon.
 
         If the file does not exist, returns None."""